requirements.txt
Pillow
matplotlib
numpy
moderngl==3.8.4
moderngl-window
opencv-python


